requirements.txt
requests
python-dotenv
LunarCalendar
